+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# Changelog
+
+## [1.1.0](https://github.com/EddieDover/mothership-crew-achievements/compare/v1.0.0...v1.1.0) (2025-10-21)
+
+
+### Features
+
+* added French language and system support ([63baf89](https://github.com/EddieDover/mothership-crew-achievements/commit/63baf89a159c98891a8338cbe0b2e5dde240c55a))
+* localize skill references in achievement system ([48b0035](https://github.com/EddieDover/mothership-crew-achievements/commit/48b0035b455d809c149a0d90d730823ebf2c0277))
 
 ## [1.0.0]